+++ ABB.Application/EntriPembayaranKass/Commands/CreatePembayaranKasCommand.cs
@@ -13,7 +13,7 @@ namespace ABB.Application.EntriPembayaranKass.Commands
     {
         public string NoVoucher { get; set; }
         public string KodeAkun { get; set; }
-        public int? TotalBayar { get; set; }
+        public decimal? TotalBayar { get; set; }
         public string FlagPembayaran { get; set; }
         public string DebetKredit { get; set; }
         public string NoNota4 { get; set; }
