+++ ABB.Domain/Entities/EntriPembayaranBankTemp.cs
@@ -14,7 +14,7 @@ namespace ABB.Domain.Entities
 
         public string NoNota4 { get; set; }
         public string KodeMataUang { get; set; }
-        public int? TotalBayar { get; set; }
+        public decimal? TotalBayar { get; set; }
        public string DebetKredit { get; set; }
         public string UserBayar { get; set; }
         public decimal? TotalDlmRupiah { get; set; }
